Unify index.js and browser.js
Note this is a backwards incompatible change!
