+++ index.js
@@ -1,11 +1,22 @@
 import { Font } from "./third_party/font.patched.js";
 import Fondue from "./src/fondue/Fondue.js";
 
-export default function loadFondue(fontPath) {
+export function fromPath(fontPath) {
 	return new Promise((resolve, reject) => {
 		const font = new Font(fontPath);
 		font.onload = () => resolve(new Fondue(font));
 		font.onerror = (e) => reject(e.detail.message);
+
 		font.src = fontPath;
+	});
+}
+
+export function fromDataBuffer(buffer, fontFilename) {
+	return new Promise((resolve, reject) => {
+		const font = new Font(fontFilename);
+		font.onload = () => resolve(new Fondue(font));
+		font.onerror = (error) => reject(error);
+
+		font.fromDataBuffer(buffer, fontFilename);
 	});
 }
