+++ rollup.config.js
@@ -6,6 +6,7 @@ export default {
 		file: "lib/index.js",
 		inlineDynamicImports: true,
 		format: "cjs",
+		exports: "named",
 	},
 	external: ["fs", "path", "util"],
 	plugins: [commonjs()],
