+++ test/support/utils.js
@@ -1,0 +1,8 @@
+export function toArrayBuffer(buffer) {
+	const ab = new ArrayBuffer(buffer.length);
+	const view = new Uint8Array(ab);
+	for (let i = 0; i < buffer.length; ++i) {
+		view[i] = buffer[i];
+	}
+	return ab;
+}
